Move isUrl function to new util module
